Properly handle both TZ-aware and naive datetime objects
datetime.timestamp() on TZ-aware objects is the simplest and safest way
to get the corresponding timestamp. So let's work with that instead.

So we now have an explicit conversion of naive datetimes into a TZ-aware
object and a single codepath to get the timestamp. time.struct_time
objects are still allowed, their code path untouched.